pyproject.toml
[build-system]
requires= ["setuptools>=42.0", "wheel"]
build-backend= "setuptools.build_meta"

[project]
name = "services"
version = "0.0.1"
description = "A quick streamlit demo with clustering models"
readme = "README.md"
authors = [
    {name = "Muhammed Amr", email = "muhammedamr57@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
]
keywords = ["machine learning", "python", "AI", "data science"]
requires-python = ">=3.10"

dependencies = [
    "Flask==3.0.3",
    "groq==0.18.0",
    "fireworks-ai==0.15.12",
    "python-dotenv==1.0.0",
    "beautifulsoup4==4.12.2",
    "uuid==1.30"
]

[project.optional-dependencies]
testing = [
    "pytest==8.3.4",
    "Sphinx==8.1.3",
    "sphinx-rtd-theme==3.0.2",
    "sphinx-autodoc-typehints==2.5.0",
    "pytest-cov==6.0.0",
    "mypy==1.13.0",
    "flake8==7.1.1",
    "tox==4.23.2",
    "black==24.10.0",
    "pytest-mock==3.14.0",
    "pytest-watcher==0.4.3",
    "autopep8==2.3.1"
]

[tool.pytest.ini_options]
addopts = "--cov=insight"
testpaths = [
    "test",
]

[tool.mypy]
mypy_path = "src"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true